fix: extract model + zod parsing
